requirements.txt
pandas
schedule
python-dotenv